REF: Move `run_logged_subprocess` to `octue.utils.processes`
