+++ octue/utils/threads.py
@@ -1,6 +1,8 @@
 import time
-from subprocess import PIPE, STDOUT, CalledProcessError, Popen
-from threading import Thread, Timer
+import warnings
+from threading import Timer
+
+from octue.utils.processes import run_logged_subprocess as moved_run_logged_sub_process
 
 
 class RepeatingTimer(Timer):
@@ -13,8 +15,7 @@ class RepeatingTimer(Timer):
 
 
 def run_logged_subprocess(command, logger, log_level="info", *args, **kwargs):
-    """Run a subprocess, sending its stdout and stderr output to the given logger. Extra `args` and `kwargs` are
-    provided to the `subprocess.Popen` instance used.
+    """The deprecated version of `octue.utils.processes.run_logged_subprocess`.
 
     :param iter(str) command: command to run
     :param logging.Logger logger: logger to use to log stdout and stderr
@@ -22,23 +23,11 @@ def run_logged_subprocess(command, logger, log_level="info", *args, **kwargs):
     :raise CalledProcessError: if the subprocess fails (i.e. if it doesn't exit with a 0 return code)
     :return subprocess.CompletedProcess:
     """
+    warnings.warn(
+        DeprecationWarning(
+            "`run_logged_subprocess` has been moved to `octue.utils.processes`. Importing it from "
+            "`octue.utils.threads` is deprecated and the ability to do so will be removed soon."
+        )
+    )
 
-    def _log_lines_from_stream(stream, logger):
-        """Log lines from the given stream.
-
-        :param io.BufferedReader stream:
-        :param logging.Logger logger:
-        :return None:
-        """
-        with stream:
-            for line in iter(stream.readline, b""):
-                getattr(logger, log_level.lower())(line.decode().strip())
-
-    process = Popen(command, stdout=PIPE, stderr=STDOUT, *args, **kwargs)
-    Thread(target=_log_lines_from_stream, args=[process.stdout, logger]).start()
-    process.wait()
-
-    if process.returncode != 0:
-        raise CalledProcessError(returncode=process.returncode, cmd=" ".join(command))
-
-    return process
+    moved_run_logged_sub_process(command, logger, log_level=log_level, *args, **kwargs)
